+++ libs/application/templates/social-insurance-administration/additional-support-for-the-elderly/src/fields/Review/review-groups/Period.tsx
@@ -1,9 +1,9 @@
 import { DataValue, ReviewGroup } from '@island.is/application/ui-components'
 import { GridColumn, GridRow } from '@island.is/island-ui/core'
 import { useLocale } from '@island.is/localization'
-import { additionalSupportForTheElderyFormMessage } from '../../../lib/messages'
+import { socialInsuranceAdministrationMessage } from '@island.is/application/templates/social-insurance-administration-core/messages'
 import { ReviewGroupProps } from './props'
-import { MONTHS } from '../../../lib/constants'
+import { MONTHS } from '@island.is/application/templates/social-insurance-administration-core/constants'
 import { getApplicationAnswers } from '../../../lib/additionalSupportForTheElderlyUtils'
 
 export const Period = ({
@@ -28,7 +28,7 @@ export const Period = ({
         <GridColumn span={['12/12', '12/12', '12/12', '5/12']}>
           <DataValue
             label={formatMessage(
-              additionalSupportForTheElderyFormMessage.info.periodTitle,
+              socialInsuranceAdministrationMessage.period.title,
             )}
             value={`${month && formatMessage(month.label)} ${selectedYear}`}
             error={hasError('period')}
