+++ libs/application/templates/social-insurance-administration/old-age-pension/src/fields/EmployersRatioMonthly/index.tsx
@@ -7,7 +7,7 @@ import { RepeaterProps, FieldBaseProps } from '@island.is/application/types'
 
 import { oldAgePensionFormMessage } from '../../lib/messages'
 import { getApplicationAnswers } from '../../lib/oldAgePensionUtils'
-import { MONTHS } from '../../lib/constants'
+import { MONTHS } from '@island.is/application/templates/social-insurance-administration-core/constants'
 import { InputController } from '@island.is/shared/form-fields'
 import { getErrorViaPath } from '@island.is/application/core'
